requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
idna==3.10
language_tool_python==2.9.4
Markdown==3.8.2
psutil==7.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
redis==6.2.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
toml==0.10.2
tqdm==4.67.1
urllib3==2.5.0
